sonar-server/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
python-dotenv>=1.0
livekit>=0.11
livekit-agents>=0.8
livekit-plugins-openai>=0.8
livekit-plugins-silero>=0.6
